pyproject.toml
[project]
name = "code-catalyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.5",
    "openai>=1.59.7",
    "packaging>=24.2",
    "phidata>=2.7.8",
    "python-frontmatter>=1.1.0",
]
